# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MERN app API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  role: string;
  createdAt: string;
  /** @nullable */
  lastLogin: string | null;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DashboardStats {
  totalUsers: number;
  activeToday: number;
  newThisWeek: number;
  accountAgeDays: number;
}

export type ActivityItemMetadata = { [key: string]: unknown };

export interface ActivityItem {
  id: string;
  type: string;
  description: string;
  timestamp: string;
  metadata?: ActivityItemMetadata;
}

export interface Note {
  id: string;
  title: string;
  content: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNoteInput {
  /** @minLength 1 */
  title: string;
  content?: string;
}

export interface UpdateNoteInput {
  /** @minLength 1 */
  title?: string;
  content?: string;
}

export interface UpdateProfileInput {
  /** @minLength 1 */
  name: string;
}

export interface ChangePasswordInput {
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
}