# lib/api-zod/src/generated/types/changePasswordInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MERN app API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChangePasswordInput {
  currentPassword: string;
  /** @minLength 6 */
  newPassword: string;
}