# lib/api-zod/src/generated/types/createNoteInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MERN app API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateNoteInput {
  /** @minLength 1 */
  title: string;
  content?: string;
}