# lib/api-zod/src/generated/types/note.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MERN app API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Note {
  id: string;
  title: string;
  content: string;
  createdAt: string;
  updatedAt: string;
}