# lib/api-zod/src/generated/types/updateNoteInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MERN app API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateNoteInput {
  /** @minLength 1 */
  title?: string;
  content?: string;
}