# lib/api-zod/src/generated/types/updateProfileInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MERN app API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProfileInput {
  /** @minLength 1 */
  name: string;
}